Added: Update blog page date field and date format (#21)
* feat: Update blog page date field to use DateTimeField

Update the `date` field in the `BlogPage` model to use `DateTimeField` instead of `DateField`. This change allows for more precise timestamp information for blog posts and a better sort order on the index page.

* feat: Update date format in blog templates

Update the date format in the `blog_index_page.html` and `blog_page.html` templates to include the time in addition to the date. This change provides more precise timestamp information for blog posts.

* feat: Update base.py to use local time zone

* Update date format in blog templates