Update author display logic and add Python files to Tailwind CSS config (#30)
* feat: Update author display logic in blog templates

Refactor the author display logic in the `blog_page.html` and `blog_index_page.html` templates to handle different scenarios. The code now checks for the presence of `page.author` and `page.owner` to display the correct author information. This change ensures consistent and accurate author attribution on the blog pages.

* refactor: Update `AuthorPage` parent page types

Update the `AuthorPage` class in `models.py` to specify the parent page types as `['AuthorIndexPage']`. This change ensures that more than one author page can be created under the author index page, maintaining the desired page structure.

* feat: Update `AuthorIndexPage` to fetch all users who have authored at least one blog post

This change modifies the `get_context` method in the `AuthorIndexPage` class to fetch all users who have authored at least one blog post. The previous implementation used a specific group (`Editors`) to filter the authors, but this change removes the dependency on the group and directly queries the `User` model. This ensures that all authors are included in the context, regardless of their group membership.

* feat: Update author display logic in blog templates

* feat: Update `tailwind.config.js` to include Python files in content

Update the `tailwind.config.js` file to include Python files (`*.py`) in the `content` array. This change ensures that the Tailwind CSS utility classes are applied to HTML, JavaScript, and Python files in the project, improving consistency and styling across all file types.

* fix(test): Fix tests after adding author to BlogPage

* feat: Add restart option to docker-compose.yml

This commit adds the `restart: always` option to the `docker-compose.yml` file. This ensures that the service will automatically restart if it crashes or if the Docker daemon restarts.

* feat: Update author display logic in blog templates

* fix: Fix bug in user_tags

* style: restyle navigation links in header

* chore: Add conditional execution for smoketests job

This commit adds a conditional execution for the `smoketests` job in the `ci.yml` workflow file. The job will now only run if the branch being pushed to is `main`. This ensures that the smoketests are only executed on the main branch, reducing unnecessary build time on other branches.

* chore: Add conditional execution for smoketests job

* chore: Update actions/checkout to v4 in ci.yml workflow

* fix(style): Fix styling on author page

* refactor: Update author index and author page templates

This commit refactors the author index and author page templates in the blog application. The changes include:
- Updating the layout and styling of author cards on the author index page
- Adding a background image to the author cards with a hover effect
- Removing the user avatar from the author cards on the author index page
- Rearranging the layout of the author page to improve readability
- Adding a clear-both class to fix floating issues on the author page

These changes improve the visual presentation and user experience of the author index and author page in the blog application.

* refactor: Add wrapper div for page body in blog_page.html