Update author display logic and templates, include Python files in Tailwind CSS, and restyle navigation links in header (#31)
* feat: Update author display logic in blog templates

Refactor the author display logic in the `blog_page.html` and `blog_index_page.html` templates to handle different scenarios. The code now checks for the presence of `page.author` and `page.owner` to display the correct author information. This change ensures consistent and accurate author attribution on the blog pages.

* feat: Update `tailwind.config.js` to include Python files in content

Update the `tailwind.config.js` file to include Python files (`*.py`) in the `content` array. This change ensures that the Tailwind CSS utility classes are applied to HTML, JavaScript, and Python files in the project, improving consistency and styling across all file types.

* feat: Update author display logic in blog templates

* style: restyle navigation links in header

* refactor: Update author index and author page templates

This commit refactors the author index and author page templates in the blog application. The changes include:
- Updating the layout and styling of author cards on the author index page
- Adding a background image to the author cards with a hover effect
- Removing the user avatar from the author cards on the author index page
- Rearranging the layout of the author page to improve readability
- Adding a clear-both class to fix floating issues on the author page

These changes improve the visual presentation and user experience of the author index and author page in the blog application.

* feat: Add recommendation for charliermarsh.ruff extension

* refactor: Update author index and author page templates

* style: Remove drop-shadow from header navigation