Added: Add "Go back to site" menu item in admin panel
This commit adds a new menu item in the admin panel called "Go back to site". It is located at the top of the menu with an icon of a home. Clicking on this menu item will redirect the user to the site's homepage. This addition improves the user experience by providing a convenient way to navigate between the admin panel and the site.