Added: Current 🇧🇷⏳ Brazilian time in header (#39)
* feat: Update requirements.txt and add brazil_time template tag

Update requirements.txt to include the pytz package and add the brazil_time.py template tag to display the current time in São Paulo, Brazil. This improves the functionality of the blog application by providing accurate time information and enhances the user experience.

* feat(wip): Add the current 🇧🇷⏳ brazilian time in header (automatic update through javascript ain't working yet)

* config: Configure ruff as default linter & formatter

* feat: Update Brazil time display in header with automatic update

Update the code in the brazil-time.html template and the brazil_time.py templatetag to improve the display of the current Brazilian time in the header of the blog application. The previous implementation used JavaScript to update the time every second, but it was not working reliably. The updated code now fetches the current time from the server using an API endpoint and updates the display every minute. This change ensures accurate and consistent time information for users visiting the blog.

* config: Configure Ruff for VS Code

* refactor: Update Dockerfile and linting configuration

Update the Dockerfile to remove the unused file and configure the linting tool, Ruff, as the default linter and formatter. This improves the development environment by removing unnecessary files and ensuring consistent code quality through automated linting and formatting.

* refactor: Add bash script for formatting code with Ruff

A new bash script, `format-back`, has been added to the `bin` directory. This script uses `docker-compose` to run the `ruff format` command, allowing for easy code formatting using the Ruff linter and formatter. This addition improves the development workflow by providing a convenient way to format code and ensure consistent code quality.

* refactor: Update Dockerfile to use correct service name in test script

* ci: Add linting step in ci.yml

* feat: Update linting step in ci.yml to use local script

The linting step in ci.yml has been updated to use the local script `bin/lint-back` instead of running it inside a Docker container. This change simplifies the CI workflow and improves efficiency by avoiding the need to spin up a Docker container for linting.

* refactor: Update Ruff configuration to ignore F403 error

The Ruff configuration file `ruff.toml` has been modified to ignore the `F403` error. This error is related to import statements and is being ignored to reduce noise in the linting output. This change improves the code quality by focusing on more critical linting issues and reducing distractions caused by false positives.

* Update src/blog/templates/blog/includes/brazil-time.html

* refactor: Update Brazil time display in header with automatic update

* refactor: Use "Normal Human Order" time 😆

* Remove specific docker compose file version

* Change  command to
Usage:  docker compose [OPTIONS] COMMAND

Docker Compose

Options:
      --ansi string                Control when to print ANSI control characters ("never"|"always"|"auto") (default
                                   "auto")
      --compatibility              Run compose in backward compatibility mode
      --env-file string            Specify an alternate environment file.
  -f, --file stringArray           Compose configuration files
      --profile stringArray        Specify a profile to enable
      --project-directory string   Specify an alternate working directory
                                   (default: the path of the Compose file)
  -p, --project-name string        Project name

Commands:
  build       Build or rebuild services
  convert     Converts the compose file to platform's canonical format
  cp          Copy files/folders between a service container and the local filesystem
  create      Creates containers for a service.
  down        Stop and remove containers, networks
  events      Receive real time events from containers.
  exec        Execute a command in a running container.
  images      List images used by the created containers
  kill        Force stop service containers.
  logs        View output from containers
  ls          List running compose projects
  pause       pause services
  port        Print the public port for a port binding.
  ps          List containers
  pull        Pull service images
  push        Push service images
  restart     Restart containers
  rm          Removes stopped service containers
  run         Run a one-off command on a service.
  start       Start services
  stop        Stop services
  top         Display the running processes
  unpause     unpause services
  up          Create and start containers

Run 'docker compose COMMAND --help' for more information on a command.

* Change docker-compose command to docker compose